refactor(hooks): organize hooks into feature-based folders
Restructure the presentation hooks directory to group related hooks by feature:
- Move auth-related hooks to `/src/presentation/hooks/auth/`
- Organize transaction hooks into nested subdirectories under `/src/presentation/hooks/transaction/` for each CRUD operation
- Update all imports across the entire codebase to match the new hook paths
- Add comprehensive unit tests for all restructured and newly created hooks

This change improves code maintainability and makes it easier to discover related hooks while preserving all existing application functionality.
