+++ src/presentation/hooks/transaction/get-transaction/use-get-transactions.test.ts
@@ -1,5 +1,5 @@
 import { renderHook, waitFor } from '@testing-library/react';
-import { useTransactions } from './use-transactions';
+import { useTransactions } from './use-get-transactions';
 import { supabase } from '@/infrastructure/supabase/supabase.client';
 import { authService } from '@/infrastructure/services/supabase-auth.service';
